# FF-3204 perf: compile regex during parsing