After benchmarking Rust core and Python SDK, I noticed that rust core performs quite badly on regex conditions. Immediate reaction: regex compilation is kinda slow and we do it on every evaluation call.

This PR moves regex compilation from evaluation phase to parsing, so it happens only once and in background (doesn't slow down evaluation calls).

## Description of changes
- This PR adds `ConditionWire` struct, which is the structure we receive from the server (what is transferred "on the wire").
- `Condition` is a new type and optimized for evaluation. There are function to convert back and forth between the two formats. In additional to parsing regex, we also prepare other conditions:
  - For comparison operator, we parse version string beforehand
  - Check that condition value matches the condition operator (e.g., `IS_NULL` expects a boolean condition value)
- Evaluation details is adjusted to store `ConditionWire`, so the change is not visible to users

## Results

Comparing against pre-optimization, regex evaluation is ~150x faster now (from ~45us down to 300ns):

```
regex-flag/get_assignment
                        time:   [295.74 ns 296.20 ns 296.71 ns]
                        thrpt:  [3.3703 Melem/s 3.3761 Melem/s 3.3813 Melem/s]
                 change:
                        time:   [-99.355% -99.354% -99.352%] (p = 0.00 < 0.05)
                        thrpt:  [+15338% +15369% +15403%]
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  7 (7.00%) high mild

regex-flag/get_assignment_details
                        time:   [1.8295 µs 1.8331 µs 1.8369 µs]
                        thrpt:  [544.40 Kelem/s 545.53 Kelem/s 546.60 Kelem/s]
                 change:
                        time:   [-96.137% -96.112% -96.093%] (p = 0.00 < 0.05)
                        thrpt:  [+2459.8% +2472.3% +2488.7%]
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
```

For python SDK, regex benchmark used to be 5x slower than native python sdk. Now it is 14x faster:
```
+----------------------------------+---------+-----------------------+-----------------------+
| Benchmark                        | 3.7.0   | 4.0.0-old             | 4.0.0                 |
+==================================+=========+=======================+=======================+
| Evaluation (sdk-test-data)       | 14.3 us | 7.10 us: 2.01x faster | 3.75 us: 3.81x faster |
+----------------------------------+---------+-----------------------+-----------------------+
| Evaluation (new-user-onboarding) | 9.27 us | 2.37 us: 3.92x faster | 361 ns: 25.67x faster |
+----------------------------------+---------+-----------------------+-----------------------+
| Evaluation (numeric-one-of)      | 9.77 us | 774 ns: 12.62x faster | 771 ns: 12.67x faster |
+----------------------------------+---------+-----------------------+-----------------------+
| Evaluation (regex-flag)          | 8.43 us | 46.1 us: 5.47x slower | 567 ns: 14.86x faster |
+----------------------------------+---------+-----------------------+-----------------------+

```